requirements.txt
python-dotenv
jinja2
bs4
requests